Updates ndk info
This changes the script generating the ndk info yaml file, in such a
ways as to avoid combinations that are not supported by the various ndk
versions.

The index view has been updated accordingly too, and now an icon for
invalid combinations is shown instead of a link, that would have
resulted in a request for a broken build.